requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==2.5.0
pymongo==3.12.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
pydantic==1.8.2
